\name{IndTestPP-package}
\alias{IndTestPP-package}
\alias{IndTestPP}
\docType{package}
\title{Tests of Independence Between  Point Processes in Time}
\description{It provides several parametric and non-parametric tests to check the independence  between two or more homogeneous and nonhomogeneous
 point processes. In addition, it provides tools for simulating  point processes in one dimension, usually time, with different types of dependence, in particular: common Poisson 
shock processes, a network of  queues with exponential arrivals, multivariate  Neyman-Scott process with dependent cluster centers, and  a Poisson process with dependent marks.}

\details{
\tabular{ll}{
Package: \tab IndTestPP\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2016-07-25\cr
License: \tab GPL (>=2)  \cr
}}

\author{Ana C. Cebrian 
Maintainer: Ana C. Cebrian  <acebrian@unizar.es>
}

\keyword{Tests of Independence Between Point Processes in Time}



