\name{IndTestPP-package}
\alias{IndTestPP-package}
\alias{IndTestPP}
\docType{package}
\title{Tests of Independence  and Analysis of Dependence between  Point Processes in Time}
\description{It provides  a  general framework to analyse dependence between point processes in time. The package
includes several parametric and non-parametric tests to check the independence  
between two or more homogeneous and nonhomogeneous point processes and some tools to characterize the dependence
between them, if it exists. In addition, it includes functions for generating and analysing dependent
point processes in time with different types of dependence.}

\details{
\tabular{ll}{
Package: \tab IndTestPP\cr
Type: \tab Package\cr
Version: \tab 3.0\cr
Date: \tab 2020-08-28\cr
License: \tab GPL (>=2)  \cr
}}

\author{Ana C. Cebrian 
Maintainer: Ana C. Cebrian  <acebrian@unizar.es>
}

\keyword{Vectors of Point Processes in Time}



